ORB-546 - L.5. - Use of 32 bits does not save gas

+++ contracts/PositionManager.sol
@@ -161,7 +161,7 @@ contract PositionManager is IPositionManager, ERC721Holder, Initializable {
 
         bytes32[] memory moduleKeys = Storage.registry.getModuleKeys();
 
-        for (uint32 i = 0; i < moduleKeys.length; i++) {
+        for (uint256 i = 0; i < moduleKeys.length; i++) {
             (address moduleAddress, , bytes32 defaultData, bool activatedByDefault) = Storage.registry.getModuleInfo(
                 moduleKeys[i]
             );
